Adding code to register the device.
Database now has a function to send device data.
UIDevice extension added to give the name of the current device.
All that should be needed now is to actually call the database function
That should correctly register the device with the server.

Now I gotta go watch a dog die. ☹️